Modify _load_npz_proatom & _get_proatom_density
_load_npz_proatom() is modified to deal with the current data in
"grid.data.proatom" directory as not all atoms contain neutral, cation
and anion information.

_get_proatom_density() first part is modified to include extra code to
deal with the data loaded in _load_npz_proatom(). In the second part of
the function, new code is include to compute proatomic density based on
the charges and coefficients reported for that atom in the “expansion”
dictionary passed for the molecule.